Fixed a few of the warnings in rwhw.
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@5995 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWorkHardware/src/rwhw/universalrobots/URCommon.hpp
@@ -268,7 +268,7 @@ public:
                         /*socket->send(boost::asio::buffer(buffer, n));*/
                         std::size_t bytesTransfered = 0;
                         bytesTransfered = boost::asio::write(*socket, boost::asio::buffer(buffer, n));
-                        if (bytesTransfered == n) {
+                        if (static_cast<int>(bytesTransfered) == n) {
                             /* Successful send */
                             RW_LOG_DEBUG("Sent all of the '" << bytesTransfered << "' bytes.");
                         } else {
